requirements.txt
fastapi==0.115.12
uvicorn==0.23.2
httpx==0.25.0
python-consul==1.1.0
prometheus_client==0.21.1
python-dotenv>=1.0.0
prometheus_fastapi_instrumentator==7.1.0
circuitbreaker==2.1.3
python-json-logger==2.0.7
loguru==0.7.2

asyncpg>=0.27.0
schedule>=1.2.0
colorama>=0.4.6
python-multipart>=0.0.6
alembic>=1.10.3

psycopg2-binary>=2.9.6
